requirements.txt
django==3.1
djangorestframework==3.11.0
django-cors-headers==3.2.1
razorpay==1.2.0
requests==2.23.0